:sparkles: 투표 API 수정 (투표 마감여부 반환값 추가 및 중복 투표 시 409 Http Status Code 리턴)
