+++ src/main/java/kr/co/antoon/common/dto/ResponseDto.java
@@ -21,4 +21,10 @@ public record ResponseDto<T>(T date) implements Serializable {
                 .status(HttpStatus.NO_CONTENT)
                 .build();
     }
+
+    public static ResponseEntity<Void> conflict() {
+        return ResponseEntity
+                .status(HttpStatus.CONFLICT)
+                .build();
+    }
 }
