+++ src/main/java/kr/co/antoon/error/dto/ErrorMessage.java
@@ -27,6 +27,7 @@ public enum ErrorMessage {
     FILE_UPLOAD_ERROR("오류로 인해 파일 업로드에 실패하였습니다."),
     NOT_EXISTS_CANDIDATE("존재하지 않는 후보입니다."),
     MAPPER_JSON_ERROR("mapper error가 발생했습니다."),
+    ALREADY_VOTE_ERROR("이미 투표했습니다"),
     ;
 
     private final String description;
