+++ src/main/java/kr/co/antoon/vote/infrastructure/VoteRepository.java
@@ -6,4 +6,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface VoteRepository extends JpaRepository<Vote, Long> {
+    Boolean existsByUserIdAndTopicId(long userId, long topicId);
 }
