+++ src/main/java/kr/co/antoon/vote/presentation/VoteController.java
@@ -2,6 +2,7 @@ package kr.co.antoon.vote.presentation;
 
 import io.swagger.annotations.Api;
 import kr.co.antoon.common.dto.ResponseDto;
+import kr.co.antoon.error.exception.common.AlreadyExistsException;
 import kr.co.antoon.oauth.config.AuthUser;
 import kr.co.antoon.oauth.dto.AuthInfo;
 import kr.co.antoon.vote.facade.VoteFacade;
@@ -21,7 +22,11 @@ public class VoteController {
             @PathVariable Long candidateId,
             @AuthUser AuthInfo info
     ) {
-        voteFacade.create(candidateId, info.userId());
-        return ResponseDto.noContent();
+        try {
+            voteFacade.create(candidateId, info.userId());
+            return ResponseDto.noContent();
+        } catch (AlreadyExistsException e) {
+            return ResponseDto.conflict();
+        }
     }
 }
